@@ -4,6 +4,7 @@
   "description": "A React component for xterm.js terminal integration.",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
+  "type": "module",
   "files": [
     "dist"
   ],
@@ -21,6 +22,7 @@
     "test:react": "node version-compatibility-tests/core/react-versions-test.js",
     "test:eslint": "node version-compatibility-tests/devtools/eslint-versions-test.js",
     "test:biome": "node version-compatibility-tests/devtools/biome-versions-test.js",
+    "compat:scenario": "node version-compatibility-tests/run-scenarios.mjs",
     "publish": "./scripts/prepare-publish.sh"
   },
   "husky": {
@@ -38,19 +40,19 @@
     "@types/node": "^24.3.0",
     "@types/react": "^19.0.0",
     "@types/react-dom": "^19.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.8.0",
-    "@typescript-eslint/parser": "^8.0.0",
+    "@typescript-eslint/eslint-plugin": "^8.44.1",
+    "@typescript-eslint/parser": "^8.44.1",
     "@vitejs/plugin-react-refresh": "^1.3.6",
-    "eslint": "^9.4.0",
+    "eslint": "^9.36.0",
     "eslint-config-prettier": "^10.0.2",
-    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-prettier": "^5.5.4",
     "husky": "^9.0.11",
     "prettier": "^3.3.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "rimraf": "^6.0.1",
     "typescript": "^5.4.5",
-    "typescript-eslint": "^8.0.0",
+    "typescript-eslint": "^8.44.1",
     "vite": "^7.1.2"
   },
   "dependencies": {
@@ -69,5 +71,12 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/PabloLION/xterm-react.git"
-  }
+  },
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    }
+  },
+  "sideEffects": false
 }
