@@ -1,13 +1,14 @@
 #!/usr/bin/env node
 
 /**
- * ESLint Ecosystem Version Compatibility Test
+ * ESLint Ecosystem Version Compatibility Test (ESM)
  * Tests different versions of ESLint and related tools with latest React & TypeScript
  */
 
-const { execSync } = require("child_process");
-const fs = require("fs");
-const path = require("path");
+import { execSync } from "node:child_process";
+import fs from "node:fs";
+import path from "node:path";
+import { fileURLToPath } from "node:url";
 
 const ESLINT_VERSIONS_TO_TEST = [
   {
@@ -16,7 +17,6 @@ const ESLINT_VERSIONS_TO_TEST = [
       eslint: "^8.57.0",
       "@typescript-eslint/parser": "^6.21.0",
       "@typescript-eslint/eslint-plugin": "^6.21.0",
-      "typescript-eslint": "^6.21.0",
     },
   },
   {
@@ -25,7 +25,6 @@ const ESLINT_VERSIONS_TO_TEST = [
       eslint: "^9.0.0",
       "@typescript-eslint/parser": "^7.18.0",
       "@typescript-eslint/eslint-plugin": "^7.18.0",
-      "typescript-eslint": "^7.18.0",
     },
   },
   {
@@ -34,7 +33,6 @@ const ESLINT_VERSIONS_TO_TEST = [
       eslint: "^9.36.0",
       "@typescript-eslint/parser": "^8.44.1",
       "@typescript-eslint/eslint-plugin": "^8.44.1",
-      "typescript-eslint": "^8.44.1",
     },
   },
 ];
@@ -50,9 +48,54 @@ const PRETTIER_VERSIONS_TO_TEST = [
 ];
 
 class ESLintVersionTester {
-  constructor() {
+  constructor(opts = {}) {
+    this.verbose = Boolean(opts.verbose);
+    this.logDir = opts.logDir || null;
     this.results = [];
     this.originalPackageJson = null;
+  }
+
+  slugify(text) {
+    return text
+      .toLowerCase()
+      .replace(/[^a-z0-9]+/g, "-")
+      .replace(/(^-|-$)/g, "");
+  }
+
+  execWithLog(cmd, options, logFile) {
+    const logPath = this.logDir
+      ? path.join(this.logDir, "eslint", logFile)
+      : null;
+    try {
+      const out = execSync(cmd, { ...options, stdio: "pipe" });
+      if (logPath) {
+        fs.mkdirSync(path.dirname(logPath), { recursive: true });
+        fs.writeFileSync(logPath, out);
+      }
+      if (this.verbose) {
+        const txt = out.toString();
+        const lines = txt.split(/\r?\n/);
+        const tail = lines.slice(-40).join("\n");
+        console.log(`[verbose] ${cmd} -> ${logFile}`);
+        if (tail.trim()) console.log(tail);
+      }
+      return true;
+    } catch (err) {
+      const stdout = err.stdout ? err.stdout.toString() : "";
+      const stderr = err.stderr ? err.stderr.toString() : "";
+      const combined = stdout + (stderr ? "\n" + stderr : "");
+      if (logPath) {
+        fs.mkdirSync(path.dirname(logPath), { recursive: true });
+        fs.writeFileSync(logPath, combined);
+      }
+      if (this.verbose) {
+        console.error(`[verbose] FAILED: ${cmd} -> ${logFile}`);
+        const lines = combined.split(/\r?\n/);
+        const tail = lines.slice(-60).join("\n");
+        if (tail.trim()) console.error(tail);
+      }
+      throw err;
+    }
   }
 
   async testAllVersions() {
@@ -120,13 +163,25 @@ class ESLintVersionTester {
       );
 
       // Install dependencies
-      console.log(`   Installing ${config.name} dependencies...`);
-      execSync("npm install --silent", { stdio: "pipe" });
+      if (this.verbose)
+        console.log(`   Installing ${config.name} dependencies...`);
+      const installCmd = this.verbose
+        ? "pnpm install"
+        : "pnpm install --silent";
+      this.execWithLog(
+        installCmd,
+        {},
+        `${this.slugify(config.name)}-install.log`
+      );
 
       // Test ESLint functionality
-      console.log(`   Testing ESLint execution...`);
+      if (this.verbose) console.log(`   Testing ESLint execution...`);
       try {
-        execSync("npm run lint:no-fix", { stdio: "pipe" });
+        this.execWithLog(
+          "pnpm run lint:no-fix",
+          {},
+          `${this.slugify(config.name)}-lint.log`
+        );
         var lintSuccess = true;
       } catch (lintError) {
         // ESLint finding issues is OK, we just want to ensure it runs
@@ -136,8 +191,12 @@ class ESLintVersionTester {
       }
 
       // Test build with new ESLint setup
-      console.log(`   Testing build compatibility...`);
-      execSync("npm run build", { stdio: "pipe" });
+      if (this.verbose) console.log(`   Testing build compatibility...`);
+      this.execWithLog(
+        "pnpm run build",
+        {},
+        `${this.slugify(config.name)}-build.log`
+      );
 
       return {
         category: "ESLint",
@@ -147,7 +206,9 @@ class ESLintVersionTester {
         lintExecutes: lintSuccess,
         buildSuccess: true,
         timestamp: new Date().toISOString(),
-        details: `ESLint ${lintSuccess ? "executes" : "has execution issues"}, build successful`,
+        details: `ESLint ${
+          lintSuccess ? "executes" : "has execution issues"
+        }, build successful`,
       };
     } catch (error) {
       return {
@@ -184,12 +245,23 @@ class ESLintVersionTester {
 
       // Install dependencies
       console.log(`   Installing ${config.name} dependencies...`);
-      execSync("npm install --silent", { stdio: "pipe" });
+      const installCmd2 = this.verbose
+        ? "pnpm install"
+        : "pnpm install --silent";
+      this.execWithLog(
+        installCmd2,
+        {},
+        `${this.slugify(config.name)}-install.log`
+      );
 
       // Test Prettier functionality
       console.log(`   Testing Prettier execution...`);
       try {
-        execSync("npx prettier --check src/", { stdio: "pipe" });
+        this.execWithLog(
+          "pnpm exec prettier --check src/",
+          {},
+          `${this.slugify(config.name)}-prettier-check.log`
+        );
         var prettierSuccess = true;
       } catch (prettierError) {
         // Prettier finding formatting issues is OK, we just want to ensure it runs
@@ -201,7 +273,11 @@ class ESLintVersionTester {
       // Test format command
       console.log(`   Testing format command...`);
       try {
-        execSync("npm run format", { stdio: "pipe" });
+        this.execWithLog(
+          "pnpm run format",
+          {},
+          `${this.slugify(config.name)}-format.log`
+        );
         var formatSuccess = true;
       } catch (formatError) {
         var formatSuccess = false;
@@ -216,7 +292,9 @@ class ESLintVersionTester {
         prettierExecutes: prettierSuccess,
         formatSuccess: formatSuccess,
         timestamp: new Date().toISOString(),
-        details: `Prettier ${prettierSuccess ? "executes" : "has execution issues"}, format ${formatSuccess ? "works" : "issues"}`,
+        details: `Prettier ${
+          prettierSuccess ? "executes" : "has execution issues"
+        }, format ${formatSuccess ? "works" : "issues"}`,
       };
     } catch (error) {
       return {
@@ -237,7 +315,10 @@ class ESLintVersionTester {
         JSON.stringify(this.originalPackageJson, null, 2)
       );
       try {
-        execSync("npm install --silent", { stdio: "pipe" });
+        const restoreCmd = this.verbose
+          ? "pnpm install"
+          : "pnpm install --silent";
+        this.execWithLog(restoreCmd, {}, `restore-install.log`);
         console.log("✅ Original dependencies restored");
       } catch (error) {
         console.error("⚠️  Warning: Failed to restore original dependencies");
@@ -246,6 +327,8 @@ class ESLintVersionTester {
   }
 
   generateReport() {
+    const __filename = fileURLToPath(import.meta.url);
+    const __dirname = path.dirname(__filename);
     const report = {
       testDate: new Date().toISOString(),
       baseVersions: {
@@ -283,10 +366,20 @@ class ESLintVersionTester {
       markdown += `## ESLint Versions\n\n`;
       eslintResults.forEach((result) => {
         markdown += `### ${result.name}\n\n`;
-        markdown += `- **Status**: ${result.success ? "✅ Compatible" : "❌ Issues found"}\n`;
-        markdown += `- **Versions**: ${JSON.stringify(result.versions, null, 2)}\n`;
-        markdown += `- **Lint Execution**: ${result.lintExecutes ? "✅ Works" : "❌ Issues"}\n`;
-        markdown += `- **Build**: ${result.buildSuccess ? "✅ Success" : "❌ Failed"}\n`;
+        markdown += `- **Status**: ${
+          result.success ? "✅ Compatible" : "❌ Issues found"
+        }\n`;
+        markdown += `- **Versions**: ${JSON.stringify(
+          result.versions,
+          null,
+          2
+        )}\n`;
+        markdown += `- **Lint Execution**: ${
+          result.lintExecutes ? "✅ Works" : "❌ Issues"
+        }\n`;
+        markdown += `- **Build**: ${
+          result.buildSuccess ? "✅ Success" : "❌ Failed"
+        }\n`;
         if (result.error) {
           markdown += `- **Error**: ${result.error}\n`;
         }
@@ -301,11 +394,17 @@ class ESLintVersionTester {
       markdown += `## Prettier Versions\n\n`;
       prettierResults.forEach((result) => {
         markdown += `### ${result.name}\n\n`;
-        markdown += `- **Status**: ${result.success ? "✅ Compatible" : "❌ Issues found"}\n`;
+        markdown += `- **Status**: ${
+          result.success ? "✅ Compatible" : "❌ Issues found"
+        }\n`;
         markdown += `- **Prettier Version**: ${result.version}\n`;
         markdown += `- **Plugin Version**: ${result.pluginVersion}\n`;
-        markdown += `- **Prettier Execution**: ${result.prettierExecutes ? "✅ Works" : "❌ Issues"}\n`;
-        markdown += `- **Format Command**: ${result.formatSuccess ? "✅ Works" : "❌ Issues"}\n`;
+        markdown += `- **Prettier Execution**: ${
+          result.prettierExecutes ? "✅ Works" : "❌ Issues"
+        }\n`;
+        markdown += `- **Format Command**: ${
+          result.formatSuccess ? "✅ Works" : "❌ Issues"
+        }\n`;
         if (result.error) {
           markdown += `- **Error**: ${result.error}\n`;
         }
@@ -325,10 +424,11 @@ class ESLintVersionTester {
   }
 }
 
+export default ESLintVersionTester;
+
 // Run if called directly
-if (require.main === module) {
+const __filename = fileURLToPath(import.meta.url);
+if (process.argv[1] && path.resolve(process.argv[1]) === __filename) {
   const tester = new ESLintVersionTester();
   tester.testAllVersions().catch(console.error);
 }
-
-module.exports = ESLintVersionTester;
